Add option to query a specific node from cluster
With the current options of the elastic exporter we had trouble gather
metrics for some nodes of the cluster that hadn't the http api enable.
This made gathering metrics for this nodes only possible by activating
the option to gather all the cluster metrics. This, for us, was not the
desired option.

To solve this problem we used the option of the elastic api to pass as
an argument the name of the node and receive the metrics for that node.

With this change we can query our nodes with the exporter and receive
the metrics for the nodes of the cluster that don't have the http api
installed.