Add filesystem IO metrics
In addition to the new IO metrics, the filesystem `data` metrics have had their variables renamed to avoid confusion. In elastic <5 the only metric within the filesystem output of nodeStats  was `data`, which is why I assume the naming was less specific. In elastic >5 the filesystem output of nodeStats contains both `data` as well as `io_stats`.