Refactor tests for cluster health collector
This adds a more robust test for the cluster health collector by abstracting the parsing of a response into a single function. Fixtures have been added for multiple elasticsearch versions to ensure compatibility. This is in preparation for a larger refactoring and will ensure that future changes maintain the desired functionality.

Signed-off-by: Joe Adams <github@joeadams.io>